requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.1
django-crispy-forms==1.9.2
django-email-verification==0.0.7
django-otp==1.0.0
django-phone-field==1.8.1
djangorestframework==3.11.1
dnspython==1.16.0
filelock==3.0.12
googlemaps==4.4.2
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.5
psycopg2-binary==2.8.6
py3-validate-email==0.2.9
python-memcached==1.59
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
validate-email==1.3
whitenoise==5.2.0
django-heroku